refactor: move components from pages/landing to Landing
