+++ pages/index.tsx
@@ -1,5 +1,5 @@
 import { Footer, SiteNav } from "@/components/Site";
-import Landing from "../components/pages/landing/Landing";
+import Landing from "@/components/Landing";
 
 export default function LandingPage() {
   return (
